리팩토링: service를 http method에 따라 분리
